Upd: redesigned Unit Edit Page
