+++ src/hooks/useUnitEditForm.js
@@ -49,14 +49,14 @@ export const useUnitEditFormSchema = Yup.object().shape({
     ddlCoolingCoilHandingId: Yup.number().required('This field is required!'),
     ddlHeatingCoilHandingId: Yup.number().required('This field is required!'),
     ddlValveTypeId: Yup.number().required('This field is required!'),
-    txbPreheatHWC_Cap: Yup.number().required('This field is required!'),
-    txbPreheatHWC_FlowRate: Yup.number().required('This field is required!'),
-    txbCoolingCWC_Cap: Yup.number().required('This field is required!'),
-    txbCoolingCWC_FlowRate: Yup.number().required('This field is required!'),
-    txbHeatingHWC_Cap: Yup.number().required('This field is required!'),
-    txbHeatingHWC_FlowRate: Yup.number().required('This field is required!'),
-    txbReheatHWC_Cap: Yup.number().required('This field is required!'),
-    txbReheatHWC_FlowRate: Yup.number().required('This field is required!'),
+    txbPreheatHWC_Cap: Yup.number(),
+    txbPreheatHWC_FlowRate: Yup.number(),
+    txbCoolingCWC_Cap: Yup.number(),
+    txbCoolingCWC_FlowRate: Yup.number(),
+    txbHeatingHWC_Cap: Yup.number(),
+    txbHeatingHWC_FlowRate: Yup.number(),
+    txbReheatHWC_Cap: Yup.number(),
+    txbReheatHWC_FlowRate: Yup.number(),
     ddlCoolingFluidTypeId: Yup.number().required('This field is required!'),
     ddlCoolingFluidConcentrationId: Yup.number().required('This field is required!'),
     txbCoolingFluidEntTemp: Yup.number().required('This field is required!'),
