+++ src/sections/unit-edit/index.js
@@ -1,4 +1,2 @@
-export { default as UnitEdit } from './UnitEdit';
-export { default as Drawing } from './Drawing';
-export { default as Layout } from './Layout';
+export { default as UnitInfo } from './UnitInfo';
 export { default as Selection } from './Selection';
